[refactor] : 파일 위치 변경 및 컴포넌트 분리
