+++ nbc-nep-front/src/components/video-conference/video-media-item/DefaultShareMediaItem.tsx
@@ -1,4 +1,4 @@
-import MetaAvatar from "../metaverse/avatar/MetaAvatar";
+import MetaAvatar from "../../metaverse/avatar/MetaAvatar";
 
 interface Props {
   avatar: string;
